refactor(cli): ♻️ redid config file structure
